# Split/Split.WebService.Client/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SplitWebService",
    "version": "1.0"
  },
  "paths": {
    "/Expense": {
      "get": {
        "tags": [
          "Expense"
        ],
        "operationId": "GetExpense",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseServiceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/GroupExpenses": {
      "get": {
        "tags": [
          "Expense"
        ],
        "operationId": "GetGroupExpenses",
        "parameters": [
          {
            "name": "groupId",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseListServiceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/UserExpenses": {
      "get": {
        "tags": [
          "Expense"
        ],
        "operationId": "GetUserExpenses",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseListServiceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/AddExpense": {
      "post": {
        "tags": [
          "Expense"
        ],
        "operationId": "AddExpense",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "groupId",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "sum",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseServiceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/DeleteExpense": {
      "delete": {
        "tags": [
          "Expense"
        ],
        "operationId": "DeleteExpense",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Debt": {
      "get": {
        "tags": [
          "Expense"
        ],
        "operationId": "GetDebt",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DebtModelServiceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/UserDebts": {
      "get": {
        "tags": [
          "Expense"
        ],
        "operationId": "GetUserDebts",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DebtModelListServiceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/ExpenseDebts": {
      "get": {
        "tags": [
          "Expense"
        ],
        "operationId": "GetExpenseDebts",
        "parameters": [
          {
            "name": "groupId",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DebtModelListServiceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/AddDebt": {
      "post": {
        "tags": [
          "Expense"
        ],
        "operationId": "AddDebt",
        "parameters": [
          {
            "name": "expenseId",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "sum",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "paid",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DebtModelServiceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/DeleteDebt": {
      "delete": {
        "tags": [
          "Expense"
        ],
        "operationId": "DeleteDebt",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/Group": {
      "get": {
        "tags": [
          "Group"
        ],
        "operationId": "GetGroup",
        "parameters": [
          {
            "name": "groupId",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupServiceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/AddGroup": {
      "post": {
        "tags": [
          "Group"
        ],
        "operationId": "AddGroup",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "adminId",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupServiceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/DeleteGroup": {
      "delete": {
        "tags": [
          "Group"
        ],
        "operationId": "DeleteGroup",
        "parameters": [
          {
            "name": "groupId",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/UpdateName": {
      "patch": {
        "tags": [
          "Group"
        ],
        "operationId": "UpdateName",
        "parameters": [
          {
            "name": "groupId",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "name",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupServiceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Members": {
      "get": {
        "tags": [
          "Group"
        ],
        "operationId": "GetGroupMembers",
        "parameters": [
          {
            "name": "groupId",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupMemberListServiceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/AddUserToGroup": {
      "post": {
        "tags": [
          "Group"
        ],
        "operationId": "AddUserToGroup",
        "parameters": [
          {
            "name": "groupId",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/DeleteGroupMember": {
      "delete": {
        "tags": [
          "Group"
        ],
        "operationId": "DeleteGroupMember",
        "parameters": [
          {
            "name": "groupId",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/UserGroups": {
      "get": {
        "tags": [
          "Group"
        ],
        "operationId": "GetUserGroups",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupListServiceResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DebtModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "expenseId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "debt": {
            "type": "number",
            "format": "double"
          },
          "paid": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "DebtModelListServiceResponse": {
        "type": "object",
        "properties": {
          "response": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DebtModel"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DebtModelServiceResponse": {
        "type": "object",
        "properties": {
          "response": {
            "allOf": [
              {
                "$ref": "#/components/schemas/DebtModel"
              }
            ],
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Expense": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "groupId": {
            "type": "string",
            "format": "uuid"
          },
          "sum": {
            "type": "number",
            "format": "double"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "ExpenseListServiceResponse": {
        "type": "object",
        "properties": {
          "response": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Expense"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ExpenseServiceResponse": {
        "type": "object",
        "properties": {
          "response": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Expense"
              }
            ],
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Group": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "admin": {
            "type": "integer",
            "format": "int32"
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GroupListServiceResponse": {
        "type": "object",
        "properties": {
          "response": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Group"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GroupMember": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "groupId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "GroupMemberListServiceResponse": {
        "type": "object",
        "properties": {
          "response": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupMember"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GroupServiceResponse": {
        "type": "object",
        "properties": {
          "response": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Group"
              }
            ],
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "login": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}